Quick Sort - pivot(middle)
The Java code implements the QuickSort algorithm, a widely used sorting algorithm with an average time complexity of O(n log n) and a worst-case time complexity of O(n^2). It recursively divides the array based on a pivot element, efficiently sorting the elements in place. The pivot selection from the middle contributes to its stability, and the sorted array is printed after the QuickSort process.